api/code/requirements.txt
requests
beautifulsoup4
python-dateutil
Flask
flask-cors
uwsgi
mysql-connector-python